requirements.txt
python-dotenv
langchain_community
langchain_openai
langchain_experimental
streamlit
pandas
plotly==5.23.0
plotly-express==0.4.1